FatJar with mvn package
