--- src/main/java/Application/TaquinApp.java
@@ -1,26 +1,0 @@
-package Application;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-import java.io.IOException;
-
-public class TaquinApp extends Application {
-    /**
-     * Our main function triggers the launch of the user interface initialized in the TaquinController.java
-     */
-    public static void main(String[] args) {
-        launch();
-    }
-
-    @Override
-    public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(TaquinApp.class.getResource("hello-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), 400, 300);
-        stage.setTitle("Taquin Solver");
-        stage.setScene(scene);
-        stage.show();
-    }
-}
