--- src/main/java/Game/Board/TaquinBoardDirection.java
@@ -1,11 +1,0 @@
-package Game.Board;
-
-/**
- * Abstraction representing the cardinal directions on the board.
- */
-public enum TaquinBoardDirection {
-    UP,
-    RIGHT,
-    DOWN,
-    LEFT
-}
