--- src/main/java/Game/Cell/CellFactory.java
@@ -1,7 +1,0 @@
-package Game.Cell;
-
-public interface CellFactory {
-    TaquinCell createTaquinCell(int id, int x, int y);
-
-    TaquinCell createEmpty(int x, int y);
-}
