+++ src/main/java/com/Application/Main.java
@@ -1,0 +1,8 @@
+package com.Application;
+
+public class Main {
+
+    public static void main(String[] args) {
+        TaquinApp.main(args);
+    }
+}
