+++ src/main/java/com/Game/Board/TaquinBoardDirection.java
@@ -1,0 +1,11 @@
+package com.Game.Board;
+
+/**
+ * Abstraction representing the cardinal directions on the board.
+ */
+public enum TaquinBoardDirection {
+    UP,
+    RIGHT,
+    DOWN,
+    LEFT
+}
