+++ src/main/java/com/Game/Cell/CellFactory.java
@@ -1,0 +1,7 @@
+package com.Game.Cell;
+
+public interface CellFactory {
+    TaquinCell createTaquinCell(int id, int x, int y);
+
+    TaquinCell createEmpty(int x, int y);
+}
