+++ src/main/java/com/Game/Cell/DefaultCellFactory.java
@@ -1,0 +1,14 @@
+package com.Game.Cell;
+
+public class DefaultCellFactory implements CellFactory {
+
+    @Override
+    public TaquinCell createTaquinCell(int id, int x, int y) {
+        return new StringRepresentationCell(id, x, y);
+    }
+
+    @Override
+    public TaquinCell createEmpty(int x, int y) {
+        return new StringRepresentationCell(x, y);
+    }
+}
