+++ src/main/java/com/Parser/Parser.java
@@ -1,0 +1,11 @@
+package com.Parser;
+
+import com.Game.Board.Board;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+
+public interface Parser {
+
+    public Board parseFile(File file) throws FileNotFoundException;
+}
