+++ src/main/java/module-info.java
@@ -2,6 +2,6 @@ module com.taquin.taquin {
     requires javafx.controls;
     requires javafx.fxml;
 
-    opens Application to javafx.fxml;
-    exports Application;
+    opens com.Application to javafx.fxml;
+    exports com.Application;
 }
