+++ src/main/resources/com/Application/hello-view.fxml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+
+<?import javafx.collections.FXCollections?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import java.lang.*?>
+<TilePane xmlns:fx="http://javafx.com/fxml/1" alignment="CENTER" xmlns="http://javafx.com/javafx/19"
+          fx:controller="com.Application.TaquinController">
+    <HBox alignment="CENTER" spacing="20.0">
+        <HBox alignment="CENTER">
+            <GridPane fx:id="boardDisplay">
+                <columnConstraints>
+                    <ColumnConstraints fillWidth="false" hgrow="SOMETIMES" minWidth="10.0" prefWidth="50.0"/>
+                </columnConstraints>
+                <rowConstraints>
+                    <RowConstraints fillHeight="false" minHeight="10.0" prefHeight="50.0" vgrow="SOMETIMES"/>
+                </rowConstraints>
+            </GridPane>
+        </HBox>
+        <VBox fx:id="buttonBox" alignment="CENTER_RIGHT" spacing="10.0">
+            <ComboBox fx:id="heuristicCombo" promptText="Choose Heuristic">
+                <items>
+                    <FXCollections fx:factory="observableArrayList">
+                        <String fx:value="Uniform Cost"/>
+                        <String fx:value="Displacement"/>
+                        <String fx:value="Manhattan Distance"/>
+                        <String fx:value="Linear Conflict + M.D."/>
+                    </FXCollections>
+                </items>
+            </ComboBox>
+            <ComboBox fx:id="algorithmCombo" promptText="Choose algorithm">
+                <items>
+                    <FXCollections fx:factory="observableArrayList">
+                        <String fx:value="A*"/>
+                        <String fx:value="IDA*"/>
+                    </FXCollections>
+                </items>
+            </ComboBox>
+            <ComboBox fx:id="logCombo" promptText="Log Progress?">
+                <items>
+                    <FXCollections fx:factory="observableArrayList">
+                        <String fx:value="No"/>
+                        <String fx:value="Yes"/>
+                    </FXCollections>
+                </items>
+            </ComboBox>
+            <HBox alignment="BASELINE_RIGHT" spacing="5.0">
+                <Label>Game Size</Label>
+                <TextField fx:id="sizeField" maxWidth="100.0" text="3"/>
+            </HBox>
+            <HBox alignment="BASELINE_RIGHT" spacing="5.0">
+                <Label>Shuffled distance from goal</Label>
+                <TextField fx:id="shuffleDepthField" maxWidth="100.0" promptText="Shuffle depth" text="50"/>
+            </HBox>
+            <Button fx:id="newGameButton" mnemonicParsing="false" onAction="#onNewGameClick" text="New game"/>
+            <Button fx:id="fileChooserButton" mnemonicParsing="false" onAction="#onFromFileClick" text="From file"/>
+            <Button fx:id="solveButton" mnemonicParsing="false" text="Solve" onAction="#onSolveClick"/>
+        </VBox>
+    </HBox>
+</TilePane>
